Moved calendar component and added it into GoogleMap component
