+++ src/app/calendar/page.tsx
@@ -1,17 +1,13 @@
 "use client";
 
 import { useState } from "react";
-import Calendar from "./components/Calendar";
-import { useStyles } from "./styles";
+import Calendar from "../map/components/Calendar";
+import { useStyles } from "../map/components/GoogleMaps/styles";
 
 const CalendarPage = () => {
   const { classes } = useStyles();
   const [date, setDate] = useState<Date | null>(new Date(Date.now()));
-  return (
-    <div className={classes.container}>
-      <Calendar date={date} setDate={setDate} />
-    </div>
-  );
+  return <Calendar date={date} setDate={setDate} />;
 };
 
 export default CalendarPage;
