--- src/app/calendar/styles.ts
@@ -1,10 +1,0 @@
-import { createStyles } from "@mantine/core";
-
-export const useStyles = createStyles(() => ({
-  container: {
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-    minHeight: "calc(100vh - 60px)",
-  },
-}));
